+++ src/main/java/com/ddang/usedauction/member/servie/AuthService.java
@@ -1,6 +1,6 @@
 package com.ddang.usedauction.member.servie;
 
-import com.ddang.usedauction.member.domain.entity.Member;
+import com.ddang.usedauction.member.domain.Member;
 import com.ddang.usedauction.member.repository.MemberRepository;
 import com.ddang.usedauction.security.jwt.TokenProvider;
 import com.ddang.usedauction.security.jwt.exception.CustomJwtException;
@@ -41,7 +41,7 @@ public class AuthService {
     TokenDto token = refreshTokenService.findTokenByEmail(email);
     String oldAccessToken = token.getAccessToken();
 
-    if (tokenProvider.isExpiredToken(oldAccessToken) &&
+    if (!tokenProvider.isExpiredToken(oldAccessToken) &&
         !tokenProvider.isExpiredToken(token.getRefreshToken())) {
       String newAccessToken = tokenProvider.reissueAccessToken(member.getEmail(),
           List.of(new SimpleGrantedAuthority(member.getRole().toString())));
@@ -67,7 +67,7 @@ public class AuthService {
     TokenDto token = refreshTokenService.findTokenByEmail(email);
 
     // accessToken 유효시간만큼 blackList 에 저장
-    refreshTokenService.setBlackList(dto.getAccessToken(), "access_token", expiration);
+    refreshTokenService.setBlackList(dto.getAccessToken(), token, expiration);
     refreshTokenService.deleteRefreshTokenByEmail(token.getEmail());
   }
 }
